+++ spring-blog/src/main/java/com/example/springblog/controller/PostController.java
@@ -1,0 +1,36 @@
+package com.example.springblog.controller;
+
+import com.example.springblog.dto.PostDto;
+import com.example.springblog.service.PostService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/posts")
+public class PostController {
+
+    private final PostService postService;
+
+    public PostController(PostService postService) {
+        this.postService = postService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> createPost(@RequestBody PostDto postDto) {
+        postService.createPost(postDto);
+        return new ResponseEntity<>(HttpStatus.CREATED);
+    }
+
+    @GetMapping("/all")
+    public ResponseEntity<List<PostDto>> showAllPosts() {
+        return new ResponseEntity<>(postService.getAllPosts(), HttpStatus.OK);
+    }
+
+    @GetMapping("/get/{id}")
+    public ResponseEntity<PostDto> getSinglePost(@PathVariable Long id) {
+        return new ResponseEntity<>(postService.findById(id), HttpStatus.OK);
+    }
+}