+++ spring-blog/src/main/java/com/example/springblog/exceptions/PostNotFoundException.java
@@ -1,0 +1,8 @@
+package com.example.springblog.exceptions;
+
+public class PostNotFoundException extends RuntimeException {
+
+    public PostNotFoundException(String message) {
+        super(message);
+    }
+}