+++ spring-blog/History.md
@@ -7,3 +7,4 @@ Steps of Implementation
 4. [Setup Spring Security](Howto.md#1---configure-spring-security)
 5. [Implement Login using JWT](Howto.md#2---implement-login-using-jwt).
 6. [Validate JWT](Howto.md#3---validate-jwt).
+7. [Manage Posts](Howto.md#4---manage-posts) (Create, Read).